llm_ai_agent_rag/requirements.txt
# LLM & AI
openai>=1.3.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-community>=0.0.1

# RAG & Vector Stores
chromadb>=0.4.0
pinecone-client>=2.2.0
faiss-cpu>=1.9.0
sentence-transformers>=2.2.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
pydantic>=2.0.0
pytest>=7.4.0
